+++ src/srctools/const.py
@@ -142,9 +142,7 @@ class GameID(Enum):
 
 
 class SurfFlags(Flag):
-    """The various `SURF_*`_ flags, indicating different attributes for faces.
-
-    .. _SURF_*: https://github.com/ValveSoftware/source-sdk-2013/blob/master/sp/src/public/bspflags.h#L80-L97
+    """The various :sdk-2013:`SURF_*<public/bspflags.h#L80-L97>` flags, indicating different attributes for faces.
     """
     NONE = 0
     LIGHT = 0x1
@@ -184,11 +182,9 @@ class SurfFlags(Flag):
 
 
 class BSPContents(Flag):
-    """The various `CONTENTS_*`_ flags, indicating different collision types.
+    """The various :sdk-2013:`CONTENTS_*<public/bspflags.h#L24-L76>` flags, indicating different collision types.
 
     This is normally for brushes, but is also used on other things like models.
-
-    .. _CONTENTS_*: https://github.com/ValveSoftware/source-sdk-2013/blob/master/sp/src/public/bspflags.h#L24-L76
     """
     EMPTY = 0
     SOLID = 0x1