+++ src/srctools/tokenizer.py
@@ -14,13 +14,10 @@ token/value pair. Lookahead is supported, accessed by the
 the current line number as data is read, letting you :pycode:`raise BaseTokenizer.error(...)` to easily
 produce an exception listing the relevant line number and filename.
 
-Character escapes match matches `utlbuffer.cpp <utlbuffer_esc_>`_ in the SDK.
+Character escapes matches :sdk-2013:`utlbuffer.cpp <tier1/utlbuffer.cpp#L57-L69>` in the SDK.
 Specifically, the following characters are escaped:
 ``\\\\n``, ``\\\\t``, ``\\\\v``, ``\\\\b``, ``\\\\r``, ``\\\\f``, ``\\\\a``, ``\\``, ``?``, ``'`` and ``"``.
 ``/`` and ``?`` are accepted as escapes, but not produced since they're unambiguous.
-
-.. _utlbuffer_esc: https://github.com/ValveSoftware/source-sdk-2013/blob/\
-0565403b153dfcde602f6f58d8f4d13483696a13/src/tier1/utlbuffer.cpp#L57-L69
 """
 from typing import TYPE_CHECKING, Final, NoReturn, Optional, Union
 from typing_extensions import Self, TypeAlias, overload